Moved the actual doc db models to the api instead of the pcl
