+++ MobileAppDocDBOfflineSyncSample.API/Models/Code.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MobileAppDocDBOfflineSyncSampleService.Models
+{
+    public class Code
+    {
+        public string System { get; set; }
+        public string Value { get; set; }
+    }
+}
