+++ MobileAppDocDBOfflineSyncSample.API/packages.config
@@ -2,11 +2,13 @@
 <packages>
   <package id="AutoMapper" version="3.3.1" targetFramework="net45" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi" version="5.2.3" targetFramework="net45" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net45" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net45" />
   <package id="Microsoft.AspNet.WebApi.OData" version="5.5.1" targetFramework="net45" />
   <package id="Microsoft.AspNet.WebApi.Owin" version="5.2.3" targetFramework="net45" />
   <package id="Microsoft.AspNet.WebApi.Tracing" version="5.2.3" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.3" targetFramework="net45" />
   <package id="Microsoft.Azure.DocumentDB" version="1.10.0" targetFramework="net45" />
   <package id="Microsoft.Azure.Mobile.Server" version="1.1.157.1" targetFramework="net45" />
   <package id="Microsoft.Azure.Mobile.Server.Authentication" version="1.1.157.1" targetFramework="net45" />
