+++ MobileAppDocDBOfflineSyncSample.Shared/MobileAppDocDBOfflineSyncSample.Shared.csproj
@@ -37,16 +37,11 @@
     <!-- A reference to the entire .NET Framework is automatically included -->
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Color.cs" />
     <Compile Include="Code.cs" />
     <Compile Include="ColorDto.cs" />
-    <Compile Include="FormulaComponent.cs" />
     <Compile Include="FormulaDto.cs" />
-    <Compile Include="Imitation.cs" />
     <Compile Include="Layer.cs" />
-    <Compile Include="PaintSystem.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Variant.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
